AssetBundles&DevOps/pyproject.toml
[tool.poetry]
name = "framework"
version = "0.0.1"
description = ""
authors = ["Apostolos"]
readme = "README.md"
packages = [{ include = "framework", from = "src" }]

 

[tool.poetry.dependencies]
python = "^3.8"
pyspark = "3.5.0"
delta = "0.4.2"
delta-spark = "3.2.0"
pyyaml = "^6.0.1"

 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"